+++ design/build.gradle.kts
@@ -40,6 +40,9 @@ dependencies {
     implementation(libs.timber)
     implementation(libs.coil.compose)
 
+    // glide
+    implementation(libs.glide)
+
     implementation(libs.androidx.core.ktx)
     implementation(libs.androidx.lifecycle.runtime.ktx)
     implementation(libs.androidx.activity.compose)
